Created Crypto helper to contain all methods, updated table in index page for better flow, and updated new coin form
