+++ app/views/cryptos/_form.html.erb
@@ -17,12 +17,12 @@
           </div>
           <div class="form-row">
             <div class="col-md-6">
-              <%= f.label :cost_per, "Cost Per" %>
+              <%= f.label :cost_per, "Paid/Coin (How much you paid per coin)" %>
               <%= f.text_field :cost_per, class: "form-control", placeholder: "0.00" %>
             </div>
             <div class="col-md-6">
-              <%= f.label :amount_owned, "Amount Owned" %>
-              <%= f.text_field :amount_owned, class: "form-control", placeholder: "0.00" %>
+              <%= f.label :amount_owned, "Total Owned" %>
+              <%= f.text_field :amount_owned, class: "form-control", placeholder: "0" %>
             </div>
           </div><br>
           <div>
